# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MIZ API
  version: 0.1.0
  description: The authoritative HTTP contract between the Rust API and MIZ clients.
servers:
  - url: /
security:
  - sessionCookie: []
paths:
...
  /api/v1/users/me/deletion-requests/current/restore:
    post:
      operationId: restoreAccount
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RestoreAccountRequest' }
      responses:
        '200':
          description: Account restored after identity verification
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountDeletionRequest' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-miz_session
    operatorSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-miz_operator_session
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    ObjectId:
      type: string
      pattern: '^[0-9A-Za-z]{22}$'
      minLength: 22
      maxLength: 22
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 timestamp normalized to UTC.
    Handle:
      type: string
      pattern: '^[0-9A-Za-z](?!.*__)[0-9A-Za-z_]{1,22}[0-9A-Za-z]$'
...
    RestoreAccountRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { $ref: '#/components/schemas/Handle' }
        password: { type: string, minLength: 1, maxLength: 128, format: password }
    AccountDeletionRequest:
      type: object
      required: [id, status, requestedAt, restoreUntil]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        status: { type: string, enum: [pending, cancelled, purging, purged, restored] }
        requestedAt: { $ref: '#/components/schemas/Timestamp' }
        restoreUntil: { $ref: '#/components/schemas/Timestamp' }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        code:
          type: string
          enum: [auth_required, csrf_failed, forbidden, internal_error, resource_not_found, version_conflict, precondition_required, problem_validation_failed, handle_conflict, handle_change_too_soon, reauthentication_required, session_limit_reached, rate_limited, invalid_credentials, idempotency_required, idempotency_conflict, content_empty, content_too_long, content_too_large, edit_window_expired, cannot_follow_self, invalid_state_transition, parent_not_visible, invalid_cursor, cursor_expired, target_not_visible, cannot_block_self, cannot_mute_self, cannot_report_own_content, report_reason_required, report_already_exists, report_not_editable, invalid_report_transition, operator_auth_required, operator_mfa_required, operator_mfa_stale, operator_permission_denied, appeal_reviewer_conflict, deletion_already_pending, deletion_not_pending, restoration_window_expired, purge_in_progress, operator_username_unavailable, appeal_already_exists]
        requestId: { type: string }